backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
psycopg2-binary>=2.9.9

# Validation and settings
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM - Claude API
anthropic>=0.18.0

# Speech-to-Text
openai-whisper>=20231117
torch>=2.0.0

# Configuration
pyyaml>=6.0

# File uploads
python-multipart>=0.0.6

# Cache (optional)
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
